feat!: don't run Celery workers in dev mode
Tutor's importing * from devstack.py[1] for the development settings,
and that means that we aren't using Celery workers at all in dev mode
(see [2]).  This makes it so they don't start in dev mode, thus saving
everyboding a significant chunk of RAM.

[1] https://github.com/overhangio/tutor/blob/master/tutor/templates/apps/openedx/settings/lms/development.py#L3
[2] https://github.com/openedx/edx-platform/blob/master/lms/envs/devstack.py#L35

To do so, we rely on Docker compose profiles[3].

[3] https://docs.docker.com/compose/profiles/

BREAKING CHANGE: the `COMPOSE_PROJECT_STARTED` hook signature had to be changed to
accomodate profile selection.